Task2.1/requirements.txt
pytest>=8.0
requests>=2.31
pydantic>=2.0
pydantic-settings>=2.0
allure-pytest>=2.13
faker>=24.0
ruff>=0.4
